Fixing the confusion between pkg/db and pkg/database with renaming
