+++ pkg/migration/migrate_test.go
@@ -1,4 +1,4 @@
-package database
+package migration
 
 import (
 	"context"
